SBP-29_Rename User table name
